Fix bug preventing 'init' from working properly.
The Jambo 'init' command was not working properly after cutting over all
built-in commands to the CommandRegistry. When 'init' (or any other command)
is run, Jambo first attempts to instantiate all the built-in commands. This
is done in the CommandRegistry. Unfortunately, the ThemeUpgrader command could
not be created without a 'jambo.json', leading to a Catch 22. This then led
to the registry erroring out and the 'init' command failing to proceed. To fix
this, we make it so that the ThemeUpgrader can be safely created before 'init'
is run.

The current pattern in CommandRegistry._initialize is okay for now. When run,
the commands already graceully handle failures like a missing 'jambo.json'.
Long term, I think the real fix here is to change our pattern. In the Command
interface, we can change all methods besides 'execute' to be static. That way,
the registry doesn't actually need to instantiate a command until it's run.

TEST=manual

Made sure that I could succesfully run a 'jambo init'.