themeimporter: Add formatters-custom.js as static file
We add formatters-custom.js so HHs can update the file without having to
override the theme.

In the future, hopefully changes like this will only have to be in the
theme with the power of our new custom commands

J=SLAP-810
TEST=manual

On a new site that's been `jambo init`ed, import the HH theme. Make sure
that the js folder exists in the top-level static folder and that it has
the formatter-custom.js file. (In my case, formatters-custom didn't
already exist yet so I tested with formatters.js).