page cmd: add locales argument (#154)
page cmd: add locales argument

Adds a locales argument that can be passed into the page command.
The locales argument takes an array of locales, e.g. es, fr, it.

The page command's describe method returns an array of locales
found in config/locale_config.json.

J=SLAP-793
TEST=manual
When I run `jambo page` with the locales arg, such as
`jambo page --name test-locales --template vertical-standard --locales es fr`,
I see that es and fr configs are created for the test-locales in page in
addition to test-locales.json.
When I run `jambo describe` in a repo that has a locale_config.json
file with es and fr locales, I see that the page command returns
es and fr as options for the locales arg.