upgrade command: return empty array for branches if no jambo config
When there is no jambo config present and we are attempting to get
a theme's branches, the method should return an empty array.

This prevents an issue of running `jambo describe` and getting an
error before `jambo init` is run in a directory.

TEST=manual
Ran a `jambo describe` in an empty directory and see that results
are returned instead of an error.