commands: all methods except execute are static
Changes all Jambo commands so that every method except `execute`
is static.

This change is so that the cli will not need to instantiate a
command unless it actually needs to execute it.

Also updates the command interface so that all methods except
execute are static.

J=SLAP-817
TEST=manual
Ran all jambo commands and saw that there were no changes in behavior.