Recognize urls that start with only 1 slash as absolute
This commit updates the isAbsoluteUrl helper to recognize
urls like "/my-img.svg" as absolute urls. On a random website,
say yext.com, that url will point to yext.com/my-img.svg no matter
what page it is on.

T=https://yextops.zendesk.com/agent/tickets/366842
J=SLAP-881
TEST=manual

test that the WHO site can specify an iconUrl that points to a static
asset

tested that a fresh site can also specify an iconUrl in a CTA
that points to a static asset, and also can directly put in
img tags with a static asset src