Correct logic for determining if custom commands should be imported. (#169)
The logic for determining if custom commands should be imported had a small
bug. When `jambo --version` was invoked, Jambo attempted to import custom
commands and register them with the CLI. This behavior was incorrect.

TEST=manual

Made sure that `--version` did not cause custom commands to be imported and
added to the registry.