Add postimport hook, remove default postimport
This commit adds the postimport hook to jambo
import, and removes the default postimport
logic (which has been moved to the theme).
We can remove the logic in this specific case,
because anybody who calls jambo import will
be importing the newest version of the theme,
which will already have the postimport logic
moved into it.

This commit allows users to specify a postimport.js
(or any file extension, e.g. postimport.sh) script
to be run after an import.

This PR also updates the postupgrade script to have
the same capability (i.e. you can now specify an
upgrade.sh, instead of upgrade.js)

Finally it adds stdout, stderr, and error handling
to the commandexecutor, which might be renamed to
something like scriptexecutor in a separate PR,
now that we have custom jambo commands.

J=SLAP-833
TEST=manual

test that I can specify a postimport.sh or postimport.js
test that stdout, stderr, and errors are logged correctly