themeimporter: fix not being able to init with theme (#182)
themeimporter: fix not being able to init with theme

Fixes the issue where a user cannot run a `jambo init` with the
theme specified.

The usage of the theme importer as a constructor was originally
incorrect in the repository scaffolder; the change is made here
to use it correctly.

TEST=manual
Created a new directory locally, `cd`-ed into it, and ran
`jambo init --theme answers-hitchhiker-theme`. Saw that the
jambo repo was initialized correctly and that the theme was
also imported into the repo.