Remove the 'themes/answers-hitchhiker-theme' prefix from override's describe. (#196)
The override command's describe method prefixed all the options with the path
to the HH Theme. This is inaccurate, the options should all be relative to the
HH Theme. This was causing nobody to be able to create an override in CBD.

This is the last change that will go into Jambo v1.10.2, so the version was
also updated.

J=SLAP-1129
TEST=manual

Ran describe on a local repo and verified the HH Theme directory prefix was
no longer present in the override options. Made sure I could copy one of the
options exactly and pass it to the override command. The expected shadow was
created. Ran describe on a repo with no Theme imported and verified that the
correct output was returned.