Only load javascript files for custom commands (#198)
Ensure that files end with '.js' before trying to load the module

This fixes a bug where jambo would crash if a .DS_Store file was present in the commands folder.

J=none
TEST=manual

Test that jambo no longer crashes if a .DS_Store file is present in the command folder. Test the `jambo --help` command, the `jambo describe` command, and the custom vertical command from the theme.