Support importing by specifying a URL (#205)
Add a --themUrl option to the import command

The theme URL is now the preferred way to import a theme. This PR deprecates specifying a theme name directly.

J=SLAP-1136
TEST=manual

Test importing by HTTPS URL, SSH URL, and by theme name. Confirm that an error is thrown if neither a URL or a theme name are supplied. Confirm that an error is thrown if a URL is not supplied and the theme isn't known by jambo. Test that if both URL and theme are supplied, the URL is used. Test an upgrade. Observe the output of `jambo import --help` and `jambo describe`