Add themeUrl param to the init command (#206)
Allow a themeUrl to be passed to the init command

This is adds the same functionality that #205 does for the import command, except this is for the init command

J=SLAP-1136
TEST=manual

Test importing by theme URL, and by theme name. Confirm that init works if neither a URL or a theme name is supplied. Confirm that an error is thrown if a URL is not supplied and the theme isn't known by jambo. Test that if both URL and theme are supplied, the URL is used. Observe the output of jambo init --help and jambo describe