remove .git folder after an upgrade, for non submodules (#208)
Previously, this logic lived in the theme's postupgrade script.
However, any theme that is imported as checked-in files would
want the .git folder removed after recloning the theme. This is because
trying to clone a git repo inside another repo does not work properly.
The inner repo's files will not be properly checked in by the outer repo.

We don't need to worry about backwards compatibility issues with older
theme upgrade scripts. The theme's upgrade script uses fs-extra's removeSync
to remove the .git folder, which does not error out if it tries to remove a folder
that does not exist.
Also fixes an import.

J=SLAP-1135
TEST=manual

test that if I try to clone a repo inside another one, cloning the
outer repo into a different folder will not copy over the inner repo's
files.

did this by
1. making a branch in the theme which I cloned another repository into
2. seeing a warning message telling from git when I commit this, saying this won't work
3. cloning the theme again, in a new folder, and trying to navigate to my new branch
4. not seeing the inner repo's files

tested upgrading as a submodule and as regular checked in files
test upgrading to a branch of the theme that no longer has its upgrade script remove the .git folder