allow async command handlers, use cwd for git ops (#211)
This commit allows command handlers to be async, so that
upcoming acceptance test infrastructure can run multiple
commands one after another in the same process.

It also updates the init and import commands to perform
git operations within the current working directory, instead
of the initial directory jambo was called in. This is needed so that
acceptance tests can use jambo in a directory different than 
the one the acceptance tests live in. It also will be useful if we
ever release a Jambo API, so that Jambo can be called programmatically
in JS.

This does not change Jambo's behavior when only one
command is run per process, which is what normally happens.

J=SLAP-992
TEST=manual

test that custom commands work, init works, import works, and build works