use npmlog with colors
This commit refactors jambo's logging to use npmlog,
which is the logger used by npm (surprise!).
It adds a magenta 'jambo' log heading to differentiate it from
npm, which uses a white on black 'npm' heading.

J=SLAP-226
TEST=manual

test a regular jambo build
test jambo build when it throws an error because there's no jambo.json
test jambo page that is missing required arguments
test the custom card command in the theme
see that all the logs look as expected, and contain the same information as before