Changed upgrade functionality (#235)
Changed upgrade functionality, so that now it reverts to old theme files when it errors out.

J=SLAP-1308
TEST=manual/automatic

Tested by running jambo upgrade on the hitchhiker theme and making sure that upgrade ran unchanged when no error occurs. Also ran with an artificial error to make sure that theme files are reverted correctly.

Created acceptance test for themeupgrader that checks to make sure that when jambo upgrade errors out, the test-theme remains unchanged.