Export types with API extractor (#255)
Add API extractor to get an API Report, a d.ts bundle, and markdown docs, as in Answers Core. Also, add more exports, documentation, and link to markdown docs in README. Use npm-force-resolutions to upgrade to a validator version that fixes a security vulnerability.

Notes:
- When reviewing, you can ignore the docs files
- Forgotten export warnings in the API Report are out of scope for this PR
- As in Answers Core, there are separate `build-ci` and `api-extractor-ci` scripts, which exit with non-zero exit codes for warning and outdated api-reports
- As in Answers Core, the automatic tests will error out if the docs are out of date

J=SLAP-1495
TEST=manual

Ran the api extractor and inspected the api report, type bundle, and markdown docs. Made changes to the docs and checked that the docs validation script worked as expected.