add i18next specific locale formatting to support chinese
Prior to this, certain kinds of chinese locales would not be translated properly.
i18next will silently refuse to translate for you if your locale has any underscores in it.
There also seems to be a small issue with i18next itself, though it doesn't matter for our use case
(I opened a new isuse for it)
https://github.com/i18next/i18next/issues/1697

J=TECHOPS-2885
TEST=manual,auto

tested I can do zh-Hant_TW translations in the theme