add github action to update doc (#303)
add a github action to automate a commit for updating docs if it's out of date. The action will run api extractor in local mode to build api report file, and the logLevel is updated to 'error' so api extractor will fail in local mode and action will exit with non-zero code

also update build command to remove lib folder before running tsc and api extractor to ensure there's no old files in lib that may generate errors in local build.

J=SLAP-1902
TEST=auto

made a small change to a file's doc and push to pr. see that a commit is made from github action to update docs.