Pin gettext-extractor to avoid a breaking change (#335)
This PR pins `gettext-extractor` to a minor version (v3.6) prior to [v3.7.1](https://github.com/lukasgeiter/gettext-extractor/releases/tag/v3.7.1) that was released with a change that breaks Jambo upgrades performed with Node 12 and npm 6. That change added support for TS v5 and removed support for older TS versions. When upgrading Jambo via the code editor, the version of `gettext-extractor` was being updated to the latest version, which would in turn upgrade to the latest version of TS (v5.4.5) in the acceptable range (4-5). But, v5.4.5 of TS requires Node 14 or higher, causing an error when using `engine-strict=true` in the `.npmrc`. Also, because the code editor by default uses Node 12 and npm 6, we can't use `overrides` in the package.json (which was released in npm 8) to pin the transitive TS dep to v4.

J=WAT-4073
TEST=manual

Tried locally pointing the Theme to my local Jambo repo and saw that install was successful with `engine-strict=true`, when it gives the same error reported in the Slack thread when trying to force install an older version of Jambo.